add older_than capability to file.rmdir exec and state modules

+++ salt/modules/file.py
@@ -4128,7 +4128,7 @@ def stats(path, hash_type=None, follow_symlinks=True):
     return ret
 
 
-def rmdir(path, recurse=False, verbose=False):
+def rmdir(path, recurse=False, verbose=False, older_than=None):
     """
     .. versionadded:: 2014.1.0
     .. versionchanged:: 3006.0
@@ -4145,6 +4145,14 @@ def rmdir(path, recurse=False, verbose=False):
     verbose
         When ``verbose`` is set to ``True``, a dictionary is returned
         which contains more information about the removal process.
+
+        .. versionadded:: 3006.0
+
+    older_than
+        When ``older_than`` is set to a number, it is used to determine the
+        **number of days** which must have passed since the last modification
+        timestamp before a directory will be allowed to be removed. Setting
+        the value to 0 is equivalent to leaving it at the default of ``None``.
 
         .. versionadded:: 3006.0
 
@@ -4165,27 +4173,41 @@ def rmdir(path, recurse=False, verbose=False):
     if not os.path.isdir(path):
         raise SaltInvocationError("A valid directory was not specified.")
 
+    if older_than:
+        now = time.time()
+        try:
+            older_than = now - (int(older_than) * 86400)
+            log.debug("Now (%s) looking for directories older than %s", now, older_than)
+        except (TypeError, ValueError) as exc:
+            older_than = 0
+            log.error("Unable to set 'older_than'. Defaulting to 0 days. (%s)", exc)
+
     if recurse:
         for root, dirs, _ in os.walk(path, topdown=False):
             for subdir in dirs:
                 subdir_path = os.path.join(root, subdir)
-                try:
-                    log.debug("Removing '%s'", subdir_path)
-                    os.rmdir(subdir_path)
-                    deleted.append(subdir_path)
-                except OSError as exc:
-                    errors.append([subdir_path, str(exc)])
-                    log.error("Could not remove '%s': %s", subdir_path, exc)
+                if (
+                    older_than and os.path.getmtime(subdir_path) < older_than
+                ) or not older_than:
+                    try:
+                        log.debug("Removing '%s'", subdir_path)
+                        os.rmdir(subdir_path)
+                        deleted.append(subdir_path)
+                    except OSError as exc:
+                        errors.append([subdir_path, str(exc)])
+                        log.error("Could not remove '%s': %s", subdir_path, exc)
         ret = not errors
 
-    try:
-        os.rmdir(path)
-        deleted.append(path)
-        ret = True if ret or not recurse else False
-    except OSError as exc:
-        ret = False
-        errors.append([path, str(exc)])
-        log.error("Could not remove '%s': %s", path, exc)
+    if (older_than and os.path.getmtime(path) < older_than) or not older_than:
+        try:
+            log.debug("Removing '%s'", path)
+            os.rmdir(path)
+            deleted.append(path)
+            ret = True if ret or not recurse else False
+        except OSError as exc:
+            ret = False
+            errors.append([path, str(exc)])
+            log.error("Could not remove '%s': %s", path, exc)
 
     if verbose:
         return {"deleted": deleted, "errors": errors, "result": ret}
